refactor(plus): decouple device/home data flow and fix TS typing issues
extract plus snapshot data composables; simplify device/home views; fix SearchCard query result typing; add PrimitiveType in task util; clean shims module declaration for attributify compatibility

本次提交核心内容：
1. Plus 解耦重构
- 新增并接入：
  - `src/composables/plus/useDeviceSnapshotData.ts`
  - `src/composables/plus/useHomeSnapshotData.ts`
- 调整：
  - `src/views/DevicePage.vue`
  - `src/views/home/HomePage.vue`
  - `src/views/plus/DevicePage.vue`
  - `src/views/plus/home/HomePage.vue`
- 目标：把设备/首页数据流与 Plus 增强逻辑拆分，降低和官方主线的冲突面。

2. 类型与稳定性修复
- `src/types/global.d.ts`
  - `SelectorSearchResult.results` 修正为数组类型（解决 `SearchCard` 的 `.map`/索引类型错误）。
- `src/utils/task.ts`
  - 补充 `PrimitiveType` 定义。
- `src/shims.d.ts`
  - 改为不依赖真实包解析的声明方式，修复 `@unocss/preset-attributify` 类型声明报错。
- `src/views/plus/DevicePage.vue`
  - 补齐回调参数类型，消除隐式 `any`。
